# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/subscriptions/free": {
      "post": {
        "operationId": "SubscriptionsController_createFree",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFreeSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Suscripción gratuita creada"
          },
          "400": {
            "description": "Parámetros inválidos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "409": {
            "description": "Ya tiene suscripción activa"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear suscripción gratuita",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/company/{companyId}": {
      "get": {
        "operationId": "SubscriptionsController_findByCompany",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de suscripciones"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Empresa no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar suscripciones por empresa",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/{id}/cancel": {
      "patch": {
        "operationId": "SubscriptionsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suscripción cancelada"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Suscripción no encontrada"
          },
          "409": {
            "description": "Suscripción ya cancelada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancelar suscripción",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/{id}/upgrade": {
      "patch": {
        "operationId": "SubscriptionsController_upgrade",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suscripción actualizada"
          },
          "400": {
            "description": "Plan inválido o no compatible"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Suscripción no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upgrade de suscripción",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/{id}/downgrade": {
      "patch": {
        "operationId": "SubscriptionsController_downgrade",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suscripción actualizada"
          },
          "400": {
            "description": "Plan inválido o no compatible"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Suscripción no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Downgrade de suscripción",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/{id}/suspend": {
      "patch": {
        "operationId": "SubscriptionsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suscripción suspendida"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Suscripción no encontrada"
          },
          "409": {
            "description": "Suscripción ya suspendida"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suspender suscripción",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/{id}/addons": {
      "post": {
        "operationId": "SubscriptionsController_addAddon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Addon agregado"
          },
          "400": {
            "description": "Parámetros inválidos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Suscripción no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agregar addon a suscripción",
        "tags": ["Subscriptions"]
      },
      "get": {
        "operationId": "SubscriptionsController_getAddons",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de addons"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Suscripción no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar addons de una suscripción",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/addons/{addonId}": {
      "patch": {
        "operationId": "SubscriptionsController_updateAddon",
        "parameters": [
          {
            "name": "addonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddonQuantityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Addon actualizado"
          },
          "400": {
            "description": "Cantidad inválida"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Addon no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar cantidad de un addon",
        "tags": ["Subscriptions"]
      },
      "get": {
        "operationId": "SubscriptionsController_getAddon",
        "parameters": [
          {
            "name": "addonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Addon encontrado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Addon no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener addon por ID",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/addons/{addonId}/delete": {
      "patch": {
        "operationId": "SubscriptionsController_removeAddon",
        "parameters": [
          {
            "name": "addonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Addon eliminado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Addon no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar addon de una suscripción",
        "tags": ["Subscriptions"]
      }
    },
...
  },
  "info": {
    "title": "API Logistics",
    "description": "Plataforma de gestión logística multi-tenant.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Ingresa tu JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateFreeSubscriptionDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "description": "Company UUID",
            "format": "uuid"
          },
          "planId": {
            "type": "string",
            "description": "Plan UUID",
            "format": "uuid"
          }
        },
        "required": ["companyId", "planId"]
      },
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "newPlanId": {
            "type": "string",
            "description": "Target plan UUID",
            "format": "uuid"
          }
        },
        "required": ["newPlanId"]
      },
      "AddAddonDto": {
        "type": "object",
        "properties": {
          "addon_type": {
            "type": "string",
            "description": "Addon type identifier",
            "example": "extra_users"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of units",
            "example": 5,
            "minimum": 1
          }
        },
        "required": ["addon_type", "quantity"]
      },
      "UpdateAddonQuantityDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "description": "New quantity",
            "example": 10,
            "minimum": 1
          }
        },
        "required": ["quantity"]
      },
...
    }
  }
}